+++ crates/jlrs/src/lib.rs
@@ -162,6 +162,12 @@
 //!   `julia_module` macro is provided to easily export functions, types, and data in
 //!   combination with the macros from the Wrap module in the JlrsCore package.
 //!
+//! - `jl-docs`
+//!
+//!   This feature can be used with `julia_module` to generate documentation items whose source
+//!   links link to the original documentation in the Rust crate. This should only be used when
+//!   generating documentation to avoid leaking absolute paths in the library.
+//!
 //! - `lto`
 //!
 //!   jlrs depends on a support library written in C, if this feature is enabled this support
@@ -1172,7 +1178,10 @@ pub(crate) unsafe fn init_jlrs(install_jlrs_core: &InstallJlrsCore, allow_overri
         init_ledger();
         Stack::init(&unrooted);
 
-        #[cfg(all(feature = "static-arrays", any(feature = "local-rt", feature = "async-rt", feature = "multi-rt")))]
+        #[cfg(all(
+            feature = "static-arrays",
+            any(feature = "local-rt", feature = "async-rt", feature = "multi-rt")
+        ))]
         data::layout::static_arrays::init_static_arrays();
     }
 }