+++ crates/jlrs_macros/src/module/ir/documentation_ir.rs
@@ -2,12 +2,17 @@ use crate::{model::ItemModel, module::model::export_name::ExportName};
 
 pub struct DocIR<'a> {
     pub doc: String,
+    pub line: usize,
     pub export_name: &'a ExportName<'a>,
 }
 
 impl<'a> DocIR<'a> {
-    fn new(doc: String, export_name: &'a ExportName) -> Self {
-        DocIR { doc, export_name }
+    fn new(doc: String, line: usize, export_name: &'a ExportName) -> Self {
+        DocIR {
+            doc,
+            line,
+            export_name,
+        }
     }
 }
 
@@ -23,26 +28,34 @@ impl<'a> DocsIR<'a> {
             match item {
                 ItemModel::Const(const_model) if const_model.documentation.is_some() => {
                     let export_name = &const_model.export_name;
-                    let doc = const_model.documentation.as_ref().unwrap().to_string();
-                    let item = DocIR::new(doc, export_name);
+                    let doc = const_model.documentation.as_ref().unwrap();
+                    let line = doc.line();
+                    let doc = doc.to_string();
+                    let item = DocIR::new(doc, line, export_name);
                     docs.push(item);
                 }
                 ItemModel::Alias(alias_model) if alias_model.documentation.is_some() => {
                     let export_name = &alias_model.export_name;
-                    let doc = alias_model.documentation.as_ref().unwrap().to_string();
-                    let item = DocIR::new(doc, export_name);
+                    let doc = alias_model.documentation.as_ref().unwrap();
+                    let line = doc.line();
+                    let doc = doc.to_string();
+                    let item = DocIR::new(doc, line, export_name);
                     docs.push(item);
                 }
                 ItemModel::Function(function_model) if function_model.documentation.is_some() => {
                     let export_name = &function_model.export_name;
-                    let doc = function_model.documentation.as_ref().unwrap().to_string();
-                    let item = DocIR::new(doc, export_name);
+                    let doc = function_model.documentation.as_ref().unwrap();
+                    let line = doc.line();
+                    let doc = doc.to_string();
+                    let item = DocIR::new(doc, line, export_name);
                     docs.push(item);
                 }
                 ItemModel::Struct(struct_model) if struct_model.documentation.is_some() => {
                     let export_name = &struct_model.export_name;
-                    let doc = struct_model.documentation.as_ref().unwrap().to_string();
-                    let item = DocIR::new(doc, export_name);
+                    let doc = struct_model.documentation.as_ref().unwrap();
+                    let line = doc.line();
+                    let doc = doc.to_string();
+                    let item = DocIR::new(doc, line, export_name);
                     docs.push(item);
                 }
                 _ => (),